Refactor acceptor for easy importing
