+++ server.go
@@ -18,6 +18,7 @@ import (
 	proxy "github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/lightninglabs/aperture/lsat"
 	"github.com/lightninglabs/lndclient"
+	"github.com/lightninglabs/pool/acceptor"
 	"github.com/lightninglabs/pool/account"
 	"github.com/lightninglabs/pool/auctioneer"
 	"github.com/lightninglabs/pool/clientdb"
@@ -72,7 +73,7 @@ type Server struct {
 	cfg             *Config
 	db              *clientdb.DB
 	fundingManager  *funding.Manager
-	sidecarAcceptor *SidecarAcceptor
+	sidecarAcceptor *acceptor.SidecarAcceptor
 	lsatStore       *lsat.FileStore
 	lndServices     *lndclient.GrpcLndServices
 	lndClient       lnrpc.LightningClient
@@ -474,7 +475,7 @@ func (s *Server) setupClient() error {
 	// Create the funding manager. The RPC server is responsible for
 	// starting/stopping it though as all that logic is currently there for
 	// the other managers as well.
-	channelAcceptor := NewChannelAcceptor(s.lndServices.Client)
+	channelAcceptor := acceptor.NewChannelAcceptor(s.lndServices.Client)
 	s.fundingManager = funding.NewManager(&funding.ManagerConfig{
 		DB:                s.db,
 		WalletKit:         s.lndServices.WalletKit,
@@ -508,9 +509,9 @@ func (s *Server) setupClient() error {
 	// create a copy of the auctioneer client configuration because the
 	// acceptor is going to overwrite some of its values.
 	clientCfgCopy := *clientCfg
-	s.sidecarAcceptor = NewSidecarAcceptor(&SidecarAcceptorConfig{
+	s.sidecarAcceptor = acceptor.NewSidecarAcceptor(&acceptor.SidecarAcceptorConfig{
 		SidecarDB:      s.db,
-		AcctDB:         &accountStore{DB: s.db},
+		AcctDB:         &acceptor.AccountStore{DB: s.db},
 		Signer:         s.lndServices.Signer,
 		Wallet:         s.lndServices.WalletKit,
 		BaseClient:     s.lndClient,
